:recycle: Moved e2e utils to separate package and renamed topic controller test.
